refactor: Move eRosPrivilege to gta/enums.hpp
