+++ src/hooks/misc/has_privilege.cpp
@@ -1,20 +1,5 @@
 #include "hooking/hooking.hpp"
-
-enum class eRosPrivilege : uint32_t
-{
-	CREATE_TICKET = 1,
-	MULTIPLAYER = 3,
-	LEADERBOARD_WRITE = 4,
-	BANNED = 7,
-	PRIVATE_MSG = 9,
-	SOCIAL_CLUB = 10,
-	CHEATER = 20,
-	PLAYED_LAST_GEN = 24,
-	SPECIAL_EDITION_CONTENT = 25,
-	MP_TEXT_COMMUNICATION = 29,
-	MP_VOICE_COMMUNICATION = 30,
-	REPORTING = 31,
-};
+#include "gta/enums.hpp"
 
 namespace big
 {
